requirements.ansible.txt
ansible==10.4.0
pywinrm==0.5.0
ngrok-api==0.12.0
